Moved FightSequence into dunwoody_disaster
