+++ dunwoody_disaster/views/fightScreen.py
@@ -14,7 +14,7 @@ import dunwoody_disaster as DD
 from dunwoody_disaster.CharacterFactory import Character
 from dunwoody_disaster.views.characterState import CharacterState
 from dunwoody_disaster.views.action_selector import ActionSelector
-from FightSequence import FightSequence
+from dunwoody_disaster.FightSequence import FightSequence
 
 
 class FightScreen(QWidget):
